Separated producer from repository
- isolate list producer as a separate request-scoped bean
- made repository application-scoped
