+++ kitchensink-ear/jboss-as-kitchensink-ear-ejb/src/main/java/org/jboss/as/quickstarts/kitchensink_ear/data/MemberRepository.java
@@ -1,64 +1,44 @@
 package org.jboss.as.quickstarts.kitchensink_ear.data;
 
-import org.jboss.as.quickstarts.kitchensink_ear.model.Member;
-import java.util.List;
-import javax.annotation.PostConstruct;
-import javax.enterprise.context.RequestScoped;
-import javax.enterprise.event.Observes;
-import javax.enterprise.event.Reception;
-
-import javax.enterprise.inject.Produces;
+import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
-import javax.inject.Named;
 import javax.persistence.EntityManager;
 import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
+import java.util.List;
 
-@RequestScoped
+import org.jboss.as.quickstarts.kitchensink_ear.model.Member;
+
+@ApplicationScoped
 public class MemberRepository {
 
-    @Inject
-    private EntityManager em;
+   @Inject
+   private EntityManager em;
 
-    private List<Member> members;
+   public Member findById(Long id) {
+      return em.find(Member.class, id);
+   }
 
-    // @Named provides access the return value via the EL variable name "members" in the UI (e.g.,
-    // Facelets or JSP view)
-    @Produces
-    @Named
-    public List<Member> getMembers() {
-        return members;
-    }
+   public Member findByEmail(String email) {
+      CriteriaBuilder cb = em.getCriteriaBuilder();
+      CriteriaQuery<Member> criteria = cb.createQuery(Member.class);
+      Root<Member> member = criteria.from(Member.class);
+      // Swap criteria statements if you would like to try out type-safe criteria queries, a new
+      // feature in JPA 2.0
+      // criteria.select(member).where(cb.equal(member.get(Member_.name), email));
+      criteria.select(member).where(cb.equal(member.get("email"), email));
+      return em.createQuery(criteria).getSingleResult();
+   }
 
-    public void onMemberListChanged(@Observes(notifyObserver = Reception.IF_EXISTS) final Member member) {
-        retrieveAllMembersOrderedByName();
-    }
-
-    public Member findById(Long id) {
-        return em.find(Member.class, id);
-    }
-
-    public Member findByEmail(String email) {
-        CriteriaBuilder cb = em.getCriteriaBuilder();
-        CriteriaQuery<Member> criteria = cb.createQuery(Member.class);
-        Root<Member> member = criteria.from(Member.class);
-        // Swap criteria statements if you would like to try out type-safe criteria queries, a new
-        // feature in JPA 2.0
-        // criteria.select(member).where(cb.equal(member.get(Member_.name), email));
-        criteria.select(member).where(cb.equal(member.get("email"), email));
-        return em.createQuery(criteria).getSingleResult();
-    }
-
-    @PostConstruct
-    public void retrieveAllMembersOrderedByName() {
-        CriteriaBuilder cb = em.getCriteriaBuilder();
-        CriteriaQuery<Member> criteria = cb.createQuery(Member.class);
-        Root<Member> member = criteria.from(Member.class);
-        // Swap criteria statements if you would like to try out type-safe criteria queries, a new
-        // feature in JPA 2.0
-        // criteria.select(member).orderBy(cb.asc(member.get(Member_.name)));
-        criteria.select(member).orderBy(cb.asc(member.get("name")));
-        members = em.createQuery(criteria).getResultList();
-    }
+   public List<Member> findAllOrderedByName() {
+      CriteriaBuilder cb = em.getCriteriaBuilder();
+      CriteriaQuery<Member> criteria = cb.createQuery(Member.class);
+      Root<Member> member = criteria.from(Member.class);
+      // Swap criteria statements if you would like to try out type-safe criteria queries, a new
+      // feature in JPA 2.0
+      // criteria.select(member).orderBy(cb.asc(member.get(Member_.name)));
+      criteria.select(member).orderBy(cb.asc(member.get("name")));
+      return em.createQuery(criteria).getResultList();
+   }
 }
