+++ kitchensink-ear/jboss-as-kitchensink-ear-web/src/main/java/org/jboss/as/quickstarts/kitchensink_ear/rest/MemberResourceRESTService.java
@@ -51,7 +51,7 @@ public class MemberResourceRESTService {
     @GET
     @Produces(MediaType.APPLICATION_JSON)
     public List<Member> listAllMembers() {
-        return repository.getMembers();
+        return repository.findAllOrderedByName();
     }
 
     @GET
