+++ kitchensink/src/main/java/org/jboss/as/quickstarts/kitchensink/data/MemberListProducer.java
@@ -1,0 +1,38 @@
+package org.jboss.as.quickstarts.kitchensink.data;
+
+import javax.annotation.PostConstruct;
+import javax.enterprise.context.RequestScoped;
+import javax.enterprise.event.Observes;
+import javax.enterprise.event.Reception;
+import javax.enterprise.inject.Produces;
+import javax.inject.Inject;
+import javax.inject.Named;
+import java.util.List;
+
+import org.jboss.as.quickstarts.kitchensink.model.Member;
+
+@RequestScoped
+public class MemberListProducer {
+
+   @Inject
+   private MemberRepository memberRepository;
+
+   private List<Member> members;
+
+   // @Named provides access the return value via the EL variable name "members" in the UI (e.g.,
+   // Facelets or JSP view)
+   @Produces
+   @Named
+   public List<Member> getMembers() {
+      return members;
+   }
+
+   public void onMemberListChanged(@Observes(notifyObserver = Reception.IF_EXISTS) final Member member) {
+      retrieveAllMembersOrderedByName();
+   }
+
+   @PostConstruct
+   public void retrieveAllMembersOrderedByName() {
+      members = memberRepository.findAllOrderedByName();
+   }
+}
